--- examples/cmd/agent_deploy/gaia_agent/requirements.txt
@@ -1,1 +1,0 @@
-../../../../mcp_servers/requirements.txt
