+++ examples/gaia/gaia_agent_runner.py
@@ -2,6 +2,8 @@ import json
 import logging
 import os
 import re
+import subprocess
+import sys
 import traceback
 from typing import AsyncGenerator
 import uuid
@@ -99,10 +101,48 @@ class CustomToolResultParserFactory(ToolResultParserFactory):
         return super().get_parser(tool_type, tool_name)
 
 
+# Module-level flag to ensure dependencies are installed only once per program run
+_install_dependencies_flag = False
+
+
 class GaiaAgentRunner:
     """
     Gaia Agent Runner
     """
+
+    def _install_dependencies(self):
+        try:
+            current_dir = os.path.dirname(os.path.abspath(__file__))
+            requirements_file = os.path.join(current_dir, "requirements.txt")
+
+            if os.path.exists(requirements_file):
+                logger.info(f"Installing dependencies from {requirements_file}")
+                subprocess.check_call(
+                    [
+                        sys.executable,
+                        "-m",
+                        "pip",
+                        "install",
+                        "-U",
+                        "-r",
+                        requirements_file,
+                    ]
+                )
+                subprocess.check_call(
+                    [
+                        sys.executable,
+                        "-m",
+                        "pip",
+                        "install",
+                        "-U",
+                        "marker-pdf==1.7.5",
+                    ]
+                )
+                logger.info("Dependencies installed successfully")
+            else:
+                logger.warning(f"Requirements file not found at {requirements_file}")
+        except Exception as e:
+            logger.error(f"Failed to install dependencies: {e}")
 
     def __init__(
         self,
@@ -114,6 +154,11 @@ class GaiaAgentRunner:
         mcp_config: dict = None,
         session_id: str = None,
     ):
+        global _install_dependencies_flag
+        if not _install_dependencies_flag:
+            self._install_dependencies()
+            _install_dependencies_flag = True
+
         self.session_id = session_id or str(uuid.uuid4())
         self.agent_config = AgentConfig(
             llm_provider=llm_provider,
