+++ examples/gaia/requirements.txt
@@ -1,1 +1,50 @@
-../../mcp_servers/requirements.txt
+opencv-python
+pyshine
+gymnasium
+pygame
+openpyxl
+xls2xlsx
+wikipedia
+duckduckgo_search
+baidusearch
+aiohttp
+requests
+loguru
+oss2
+
+fastapi
+pydantic
+selenium
+frontend
+tools
+PyPDF2
+html2text
+xmltodict
+docx2markdown
+python-pptx
+faiss-cpu==1.9.0
+pycryptodome
+tabulate
+flask
+fitz
+terminal_controller # mcp server: https://github.com/GongRzhe/terminal-controller-mcp
+mcp_server_calculator # mcp server: https://github.com/githejie/mcp-server-calculator
+arxiv
+chess
+pytesseract
+youtube-transcript-api
+python-magic
+waybackpy
+markdown
+browser-use
+# marker-pdf==1.7.5
+pathvalidate
+pdfminer.six
+puremagic
+pydub
+SpeechRecognition
+html2text
+pre_commit
+smolagents
+
+ffmpeg-python
